[Composable API] Move test models to common file
[ghstack-poisoned]
